Refactor
Rimosso InterfaceConfiguration.java e InfrastructureConfiguration.java

Rimossi gli oggetti non in uso nel test del database e relative interfacce repository

Rinominato ClienteDTO in ClienteEntity

testato, tutto ok
