+++ src/main/java/com/example/gestionecomanda/Domain/GestionePrioritaOrdini.java
@@ -1,8 +1,7 @@
 package com.example.gestionecomanda.Domain;
 
-import com.example.gestionecomanda.Domain.DTO.ClienteEntity;
+import com.example.gestionecomanda.Domain.Entity.ClienteEntity;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 @Service
