+++ src/main/java/com/example/gestionecomanda/Domain/Repository/ClienteRepository.java
@@ -1,6 +1,6 @@
 package com.example.gestionecomanda.Domain.Repository;
 
-import com.example.gestionecomanda.Domain.DTO.ClienteEntity;
+import com.example.gestionecomanda.Domain.Entity.ClienteEntity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ClienteRepository extends CrudRepository<ClienteEntity, String> {
