+++ src/main/java/com/example/gestionecomanda/Domain/TestPort.java
@@ -1,6 +1,6 @@
 package com.example.gestionecomanda.Domain;
 
-import com.example.gestionecomanda.Domain.DTO.ClienteEntity;
+import com.example.gestionecomanda.Domain.Entity.ClienteEntity;
 
 public interface TestPort {
